added hnsserverip.site90 from 000webhost
also included 000webhost directory listing script
hnsserverip was used to retrieve my computer's ip address from a
document on pastebin. this document would be updated frequently from a
script running on my computer. the results would be in xml format so i
would get the paste containing the ip address by finding the tag with
the title "HnS Server IP" and it would echo the contents of that paste.